Add legacy stream writer
