+++ python/pyrogue/utilities/fileio/_LegacyStreamWriter.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python
+#-----------------------------------------------------------------------------
+# Title      : PyRogue FileIO - Stream Writer
+#-----------------------------------------------------------------------------
+# File       : pyrogue/utilities/filio/_StreamWriter.py
+# Created    : 2016-09-29
+#-----------------------------------------------------------------------------
+# Description:
+# Module for writing stream data.
+#-----------------------------------------------------------------------------
+# This file is part of the rogue software platform. It is subject to 
+# the license terms in the LICENSE.txt file found in the top-level directory 
+# of this distribution and at: 
+#    https://confluence.slac.stanford.edu/display/ppareg/LICENSE.html. 
+# No part of the rogue software platform, including this file, may be 
+# copied, modified, propagated, or distributed except according to the terms 
+# contained in the LICENSE.txt file.
+#-----------------------------------------------------------------------------
+import rogue.utilities
+import rogue.utilities.fileio
+import pyrogue
+import pyrogue.utilities.fileio as puf
+import rogue
+
+class LegacyStreamWriter(puf.StreamWriter):
+    def __init__(self, *, configEn=False, **kwargs):
+        pyrogue.DataWriter.__init__(self, **kwargs)
+        self._writer   = rogue.utilities.fileio.LegacyStreamWriter()
+        self._configEn = configEn
+
+    def getDataChannel(self):
+        return self._writer.getDataChannel()
+
+    def getYamlChannel(self):
+        return self._writer.getYamlChannel()
+
