+++ python/pyrogue/utilities/fileio/__init__.py
@@ -10,4 +10,5 @@
 from pyrogue.utilities.fileio._StreamWriter       import *
 from pyrogue.utilities.fileio._StreamReader       import *
 from pyrogue.utilities.fileio._FileReader         import *
+from pyrogue.utilities.fileio._LegacyStreamWriter import *
